moved logic into mail-service
